workin on login password comparing
