+++ server/passport/local-login.js
@@ -1,5 +1,5 @@
 const jwt = require('jsonwebtoken');
-const User = require('mongoose').model('User');
+const db = require('../../database/index.js')
 const PassportLocalStrategy = require('passport-local').Strategy;
 const config = require('../../config');
 
@@ -8,49 +8,11 @@ const config = require('../../config');
  * Return the Passport Local Strategy object.
  */
 module.exports = new PassportLocalStrategy({
-  usernameField: 'username',
-  passwordField: 'password',
   session: false,
   passReqToCallback: true
 }, (req, username, password, done) => {
-  const userData = {
-    username: username.trim(),
-    password: password.trim()
-  };
+  
 
   // find a user by name
-  return User.findOne({ username: userData.username }, (err, user) => {
-    if (err) { return done(err); }
-
-    if (!user) {
-      const error = new Error('Incorrect username or password');
-      error.name = 'IncorrectCredentialsError';
-
-      return done(error);
-    }
-
-    // check if a hashed user's password is equal to a value saved in the database
-    return user.comparePassword(userData.password, (passwordErr, isMatch) => {
-      if (err) { return done(err); }
-
-      if (!isMatch) {
-        const error = new Error('Incorrect username or password');
-        error.name = 'IncorrectCredentialsError';
-
-        return done(error);
-      }
-
-      const payload = {
-        sub: user._id
-      };
-
-      // create a token string
-      const token = jwt.sign(payload, config.jwtSecret);
-      const data = {
-        name: user.name
-      };
-
-      return done(null, token, data);
-    });
-  });
+  return db.passportValidate(username, password);
 });
