+++ server/routes/auth.js
@@ -79,24 +79,28 @@ router.post('/signup', (req, res, next) => {
     });
   }
   console.log(req.body, 'request body signup');
-  console.log(req.body.name);
   const { name, password } = req.body;
+  
   db.findUser(name, (err, user) => {
     if (err) {
       console.log('User.js post error: ', err)
     } else if (user) {
-      
-       message =`Sorry, already a user with the username: ${name}`;
-      
-    }
+   
+       return res.status(418).json({
+        success: true,
+        message: `Sorry, already a user with the username: ${name}`,
+        errors: validationResult.errors
+    });
+
+    }else{
+    db.saveUser(name, password);
+    return res.status(200).json({
+      success: true,
+      message: validationResult.message,
+      errors: validationResult.errors
   });
-  db.saveUser(name, password);
 
-  return res.status(200).json({
-    success: true,
-    message: validationResult.message,
-    errors: validationResult.errors
-
+    }
 });
   });
 
@@ -109,9 +113,30 @@ router.post('/login', (req, res, next) => {
       errors: validationResult.errors
     });
   }
+  const { name, password } = req.body;
+  db.findUser(name, (err, user) => {
+    if (err) {
+      console.error(err);
+    } else if (user) {
+      db.comparePassword(user.password).then((res =>{
+        if(!!res){
+          return res.status(200).json({
+            success: true,
+            message: validationResult.message,
+            errors: validationResult.errors
+        });
+        }
+      }));
+    }else{
+      return res.status(400).json({
+        success: false,
+        message: validationResult.message,
+        errors: validationResult.errors
+      });
 
-
-  return passport.authenticate('local-login', (err, token, userData) => {
+    }
+});
+   passport.authenticate('local-login', (err, token, userData) => {
     if (err) {
       if (err.name === 'IncorrectCredentialsError') {
         return res.status(400).json({
